# workspace/hooksentinel/specs/openapi.yaml
openapi: 3.1.0
info:
  title: HookSentinel Webhook Gateway API
  description: Hochperformantes, fehlertolerantes Webhook-Ingestion- & Relay-Gateway mit HMAC-SHA256-Prüfung, Idempotenz und DLQ.
  version: 1.0.0
servers:
  - url: http://localhost:8000/api/v1
    description: Lokale Instanz

paths:
  /webhook/{source_slug}:
    post:
      summary: Universelle Webhook-Ingestion
      description: Nimmt Webhook-Payloads asynchron entgegen, prüft HMAC-SHA256 und Idempotency-Keys und quittiert unverzüglich mit 202 Accepted.
      operationId: ingestWebhook
      parameters:
        - name: source_slug
          in: path
          required: true
          description: Eindeutiger Slug der Webhook-Quelle
          schema:
            type: string
            example: stripe-prod
        - name: X-Signature
          in: header
          required: false
          description: HMAC-SHA256 Signatur (Hexadezimal, z.B. sha256=...)
          schema:
            type: string
            example: sha256=7d4bf03a9856a908be54ec47eb084dc5ab5003bf3abeb9a7f34045f948bf8a8b
        - name: X-Hub-Signature-256
          in: header
          required: false
          description: GitHub-kompatibler Signatur-Header
          schema:
            type: string
        - name: X-Idempotency-Key
          in: header
          required: false
          description: Eindeutiger Idempotenzschlüssel zur Deduplizierung
          schema:
            type: string
            example: evt_1OabcXYZ987123
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: true
            example:
              event_type: payment.succeeded
              data:
                amount: 4900
                currency: eur
      responses:
        "202":
          description: Webhook akzeptiert und zur asynchronen Auslieferung eingereiht
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WebhookIngestResponse"
        "400":
          description: Ungültige Payload oder fehlende/inkorrekte Header
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: HMAC-Signaturprüfung fehlgeschlagen
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Idempotency Conflict (Webhook wurde bereits verarbeitet)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Rate Limit überschritten
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Interner Gateway-Fehler
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /dlq:
    get:
      summary: Dead-Letter-Queue Einträge abrufen
      operationId: listDlqEvents
      responses:
        "200":
          description: Liste der fehlgeschlagenen Webhook-Events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DLQEvent"

  /dlq/{event_id}/replay:
    post:
      summary: Manuelles Replay eines DLQ-Events
      operationId: replayDlqEvent
      parameters:
        - name: event_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Event erfolgreich auf PENDING zurückgesetzt
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ReplayResponse"

components:
  schemas:
    WebhookIngestResponse:
      type: object
      required:
        - status
        - event_id
        - source_slug
      properties:
        status:
          type: string
          example: accepted
        event_id:
          type: string
          example: evt_01HW8P9B2KQ91N
        source_slug:
          type: string
          example: stripe-prod
        idempotency_key:
          type: string
          nullable: true
          example: evt_1OabcXYZ987123

    DLQEvent:
      type: object
      required:
        - id
        - source_slug
        - failure_reason
        - retry_count
      properties:
        id:
          type: string
        source_slug:
          type: string
        failure_reason:
          type: string
          example: "HTTP 503 Service Unavailable"
        retry_count:
          type: integer
          example: 3
        created_at:
          type: string
          format: date-time

    ReplayResponse:
      type: object
      required:
        - replayed
        - event_id
        - new_status
      properties:
        replayed:
          type: boolean
          example: true
        event_id:
          type: string
        new_status:
          type: string
          example: pending

    ErrorResponse:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          example: invalid_signature
        message:
          type: string
          example: "HMAC-SHA256 signature verification failed"